Refactor tool interfaces and types for improved organization
