+++ packages/twenty-server/src/engine/core-modules/tool/types/tool-context.type.ts
@@ -1,0 +1,6 @@
+export type ToolContext = {
+  workspaceId: string;
+  userId?: string;
+  roleId?: string;
+  [key: string]: unknown;
+};
