+++ packages/twenty-server/src/engine/core-modules/tool/types/tool-input.type.ts
@@ -1,0 +1,6 @@
+import { ToolContext } from 'src/engine/core-modules/tool/types/tool-context.type';
+
+export type ToolInput = {
+  parameters: Record<string, unknown>;
+  context: ToolContext;
+};
