+++ packages/twenty-server/src/engine/core-modules/tool/types/tool-output.type.ts
@@ -1,0 +1,4 @@
+export type ToolOutput = {
+  result?: unknown;
+  error?: string;
+};
